[WIP] 에러 수정 및 병합 중

+++ server/src/main/java/com/example/seb_main_project/post/entity/Post.java
@@ -5,9 +5,6 @@ import com.example.seb_main_project.audit.Auditable;
 import com.example.seb_main_project.comment.entity.Comment;
 import com.example.seb_main_project.member.entity.Member;
 import com.example.seb_main_project.postlike.entity.PostLike;
-import com.example.seb_main_project.comment.entity.Comment;
-import com.example.seb_main_project.like.postlike.entity.PostLike;
-import com.example.seb_main_project.member.entity.Member;
 import lombok.*;
 
 import javax.persistence.*;
@@ -21,7 +18,6 @@ import java.util.List;
 @NoArgsConstructor
 @Entity
 public class Post extends Auditable {
-
     //< 기본 칼럼 설정 >
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -50,7 +46,7 @@ public class Post extends Auditable {
     @JoinColumn(name = "MEMBER_ID")
     private Member member;
     @Column
-    private long likeCount;
+    private Integer likeCount;
 
     @OneToMany(mappedBy = "post", cascade = CascadeType.ALL)
     private List<Comment> comments = new ArrayList<>();
